requirements.txt
pdf2image==1.10.0
pytesseract==0.3.0
opencv_python==4.1.2.30
numpy==1.19.4
pandas==0.25.3
Pillow==7.0.0
fastapi[all]
pytest==6.2.2
preprocess
GCVocr
spellcorrection
opencv-python
matplotlib